e-sink.el: fix startup time
The first two options that run-at-time() take are: when to execute
first, and how often to repeat.  Give the startup a different (and
much smaller value) than the repeat time.  While at it, use the the
repeat parameter of run-at-time(), instead of setting up a repeat by
hand in e-sink-insert-from-temp().