e-sink.el: guard e-sink-finish() against missing buffer
If we kill the buffer that e-sink is outputting to prematurely,
e-sink.pl exits with an ugly error:

  *ERROR*: No buffer named *e-sink*
  system call with parameters emacsclient --no-wait
  --eval (e-sink-finish "" ) failed:

  Inappropriate ioctl for device at /usr/bin/e-sink.pl line 47,
  <STDIN> line 483693.

We can prevent this and exit cleanly by guarding the body of
e-sink-finish() with a check to see if the buffer exists.  If it
doesn't, there's nothing to do: our previous patch makes sure that
e-sink-insert-from-temp() cancels the timer.